chore: update to Rust 1.95 (latest stable) and resolver v3
- rust-version: 1.85 → 1.95 (aligns with current stable toolchain)
- resolver: 2 → 3 (edition 2024 default; per-crate feature unification)
- Add rust-toolchain.toml pinning channel 1.95 with clippy + rustfmt
- Update README.md MSRV references (1.85 → 1.95)
- Clippy --fix: remove unnecessary closures, parentheses, collapsible ifs
- Fix empty line after doc comment in jwt.rs
- Fix non-binding let on future in masque.rs (let _ = → drop())
- All tests pass, clippy warnings reduced from ~15 to 2 (intentional)

+++ crates/voip-client/src/probe.rs
@@ -81,7 +81,7 @@ impl PortPredictionProber {
         let range_size = (port_end as u32) - (port_start as u32) + 1;
         let max_probes = self.config.prediction_max_probes;
 
-        if range_size > max_probes as u32 {
+        if range_size > max_probes {
             warn!(
                 range_size,
                 max_probes,
